+++ src/models/actions.py
@@ -5,7 +5,7 @@ from numpydantic import NDArray, Shape
 from pydantic import RootModel
 
 
-class DirectionVector(RootModel[NDArray[Shape["6"], np.float64]]):
+class DirectionVector(RootModel[NDArray[Shape["6"], np.float32]]):
     """A direction vector for ROV movement."""
